+++ NEWS.md
@@ -36,6 +36,7 @@
 - Fixed package completion tooltips #12147
 - Fixed setting rsession log level using command-line argument or logging.conf #12557
 - Fixed issue that allowed users to overwrite their home directory in server mode #12653
+- Fixed "Check for updates" incorrectly reports that there are no updates (rstudio-pro #3388)
 
 ### Accessibility Improvements